requirements.txt
djangorestframework
asgiref==3.2.10
Django==3.0.8
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-js-asset==1.2.2
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
